+++ src/libraries/Microsoft.Extensions.Diagnostics/ref/Microsoft.Extensions.Diagnostics.cs
@@ -1,7 +1,7 @@
 // Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-namespace Microsoft.Extensions.Metrics
+namespace Microsoft.Extensions.Diagnostics.Metrics
 {
     public static class MetricsServiceExtensions
     {
@@ -9,6 +9,6 @@ namespace Microsoft.Extensions.Metrics
     }
     public static class MeterFactoryExtensions
     {
-        public static System.Diagnostics.Metrics.Meter Create(this Microsoft.Extensions.Metrics.IMeterFactory meterFactory, string name, string? version = null, System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string, object?>>? tags = null,  object? scope = null) { return null!; }
+        public static System.Diagnostics.Metrics.Meter Create(this Microsoft.Extensions.Diagnostics.Metrics.IMeterFactory meterFactory, string name, string? version = null, System.Collections.Generic.IEnumerable<System.Collections.Generic.KeyValuePair<string, object?>>? tags = null,  object? scope = null) { return null!; }
     }
 }
